implementing io.py changes
I implemented the changes Tom made to io.py into your model.ipynb
driver. It makes it so that instead of calling on a directory that’s
specific computer (which makes it throw an error for other users), it
specifies the directory of the file you’re trying to read based on the
location of io.py, which is the same for all users.

Also, looks like you’re still working on the model.genorator function,
because it’s written in your model.ipynb but that function doesn’t yet
exist. I put the io.py implementation into that cell, but it doesn’t
work yet because of the aforementioned issue.

Good work!